Time: 25 ms (56.92%), Space: 19.7 MB (81.79%) - LeetHub

+++ 42-trapping-rain-water/42-trapping-rain-water.cpp
@@ -2,10 +2,10 @@ class Solution {
 public:
     int trap(vector<int>& height) {
         
-        long long int n = height.size();
-        long long int left = 0, right = n-1;
-        long long int res = 0;
-        long long int maxLeft = 0, maxRight = 0;
+        int n = height.size();
+        int left = 0, right = n-1;
+        int res = 0;
+        int maxLeft = 0, maxRight = 0;
         
         while(left <= right)
         {
@@ -18,10 +18,9 @@ public:
                 
                 left++;
             }
-            
             else
             {
-                if(height[right] >= maxRight)                    
+                if(height[right] >= maxRight)
                     maxRight = height[right];
                 else
                     res += maxRight - height[right];
